Moved crawlers over to run under nosetests
