+++ crawler/tests/test_distributed_scheduler.py
@@ -1,69 +1,12 @@
 '''
 Offline tests
 '''
-
-import unittest
 from unittest import TestCase
 import mock
 from mock import MagicMock
-
-import sys
-from os import path
-sys.path.append(path.dirname(path.dirname(path.abspath(__file__))))
-
-from crawling.redis_dupefilter import RFPDupeFilter
-from crawling.redis_retry_middleware import RedisRetryMiddleware
 from crawling.distributed_scheduler import DistributedScheduler
-from crawling.spiders.link_spider import LinkSpider
-from crawling.spiders.redis_spider import RedisSpider
-
-from crawling.log_retry_middleware import LogRetryMiddleware
-
-from scrapy.http import HtmlResponse
 from scrapy.http import Request
-from crawling.items import RawResponseItem
-
 from scutils.redis_throttled_queue import RedisThrottledQueue
-
-
-class TestRedisDupefilter(TestCase):
-
-    def setUp(self):
-        self.dupe = RFPDupeFilter(MagicMock(), 'key', 1)
-
-    def test_dupe_filter(self):
-        req = Request('http://example.com')
-        req.meta['crawlid'] = "abc123"
-
-        self.dupe.server.expire = MagicMock()
-
-        # successfully added
-        self.dupe.server.sadd = MagicMock(return_value=1)
-        self.assertFalse(self.dupe.request_seen(req))
-
-        # unsuccessfully added
-        self.dupe.server.sadd = MagicMock(return_value=0)
-        self.assertTrue(self.dupe.request_seen(req))
-
-
-class TestRedisRetryMiddleware(TestCase):
-
-    def setUp(self):
-        self.retry = RedisRetryMiddleware(MagicMock())
-        self.retry.max_retry_times = 1
-
-    def test_retries(self):
-        req = Request('http://example.com')
-        req.meta['retry_times'] = 0
-        req.meta['priority'] = 70
-
-        # number of retries less than max
-        out = self.retry._retry(req, 'stuff', MagicMock())
-        self.assertEqual(req.url, out.url)
-        self.assertEqual(60, out.meta['priority'])
-
-        # over max
-        self.assertEqual(self.retry._retry(out, 'stuff', MagicMock()), None)
 
 
 class ThrottleMixin(object):
@@ -433,222 +376,3 @@ class TestDistributedSchedulerParseCookie(ThrottleMixin, TestCase):
         }
 
         self.assertEqual(result, self.scheduler.parse_cookie(cookie))
-
-class OverrideSpider(RedisSpider):
-        name = 'OverrideSpider'
-
-class TestRedisSpider(TestCase):
-
-    def setUp(self):
-        self.spider = OverrideSpider()
-        self.spider.settings = {}
-        self.spider.redis_conn = MagicMock()
-        self.spider._logger = MagicMock()
-        self.spider.settings['STATS_CYCLE'] = 5
-        self.spider.settings['STATS_TIMES'] = []
-        self.spider.settings['STATS_RESPONSE_CODES'] = []
-        self.spider._get_hostname = MagicMock(return_value='host1')
-
-    def test_load_stats_codes(self):
-        self.spider.stats_dict = {}
-
-        # test nothing
-        self.spider._setup_stats_status_codes()
-        self.assertEquals(self.spider.stats_dict['status_codes'].keys(), [])
-
-        # test status codes only
-        self.spider.settings['STATS_RESPONSE_CODES'] = [200, 403]
-        self.spider._setup_stats_status_codes()
-        self.assertEquals(
-            sorted(self.spider.stats_dict['status_codes'].keys()),
-            sorted([200, 403]))
-        self.assertEqual(self.spider.stats_dict['status_codes'][200].keys(),
-                         ['lifetime'])
-        self.assertEqual(self.spider.stats_dict['status_codes'][403].keys(),
-                         ['lifetime'])
-
-        # test good/bad rolling stats
-        self.spider.stats_dict = {}
-        self.spider.settings['STATS_TIMES'] = [
-            'SECONDS_15_MINUTE',
-            'SECONDS_1_HOUR',
-            'SECONDS_DUMB',
-        ]
-        good = [
-            'lifetime',  # for totals, not DUMB
-            900,
-            3600,
-        ]
-
-        # check that both keys are set up
-        self.spider._setup_stats_status_codes()
-        self.assertEquals(
-            sorted(self.spider.stats_dict['status_codes'][200].keys()),
-            sorted(good))
-        self.assertEquals(
-            sorted(self.spider.stats_dict['status_codes'][403].keys()),
-            sorted(good))
-
-        k1 = 'stats:crawler:host1:OverrideSpider:200'
-        k2 = 'stats:crawler:host1:OverrideSpider:403'
-
-        for time_key in self.spider.stats_dict['status_codes'][200]:
-            if time_key == 0:
-                self.assertEquals(
-                    self.spider.stats_dict['status_codes'][200][0].key,
-                    '{k}:lifetime'.format(k=k1)
-                    )
-            else:
-                self.assertEquals(
-                    self.spider.stats_dict['status_codes'][200][time_key].key,
-                    '{k}:{t}'.format(k=k1, t=time_key)
-                    )
-
-        for time_key in self.spider.stats_dict['status_codes'][403]:
-            if time_key == 0:
-                self.assertEquals(
-                    self.spider.stats_dict['status_codes'][403][0].key,
-                    '{k}:lifetime'.format(k=k2)
-                    )
-            else:
-                self.assertEquals(
-                    self.spider.stats_dict['status_codes'][403][time_key].key,
-                    '{k}:{t}'.format(k=k2, t=time_key)
-                    )
-
-class TestLogRetryMiddlewareStats(TestCase):
-
-    @mock.patch('crawling.log_retry_middleware.LogRetryMiddleware' \
-                '.setup')
-    def setUp(self, s):
-        self.lrm = LogRetryMiddleware(MagicMock())
-        self.lrm.settings = {}
-        self.lrm.name = 'OverrideSpider'
-        self.lrm.redis_conn = MagicMock()
-        self.lrm.logger = MagicMock()
-        self.lrm.settings['STATS_CYCLE'] = 5
-        self.lrm.settings['STATS_TIMES'] = []
-        self.lrm._get_hostname = MagicMock(return_value='host1')
-
-
-    def test_lrm_stats_setup(self):
-        self.lrm.stats_dict = {}
-
-        # test nothing
-        self.lrm._setup_stats_status_codes()
-        self.assertEquals(self.lrm.stats_dict.keys(), ['lifetime'])
-
-        # test good/bad rolling stats
-        self.lrm.stats_dict = {}
-        self.lrm.settings['STATS_TIMES'] = [
-            'SECONDS_15_MINUTE',
-            'SECONDS_1_HOUR',
-            'SECONDS_DUMB',
-        ]
-        good = [
-            'lifetime',  # for totals, not DUMB
-            900,
-            3600,
-        ]
-
-        # check that both keys are set up
-        self.lrm._setup_stats_status_codes()
-        self.assertEquals(
-            sorted(self.lrm.stats_dict.keys()),
-            sorted(good))
-
-        k1 = 'stats:crawler:host1:OverrideSpider:504'
-
-        for time_key in self.lrm.stats_dict:
-            if time_key == 0:
-                self.assertEquals(
-                    self.lrm.stats_dict[0].key,
-                    '{k}:lifetime'.format(k=k1)
-                    )
-            else:
-                self.assertEquals(
-                    self.lrm.stats_dict[time_key].key,
-                    '{k}:{t}'.format(k=k1, t=time_key)
-                    )
-
-
-class TestLinkSpider(TestCase):
-
-    def setUp(self):
-        self.spider = LinkSpider()
-        self.spider._logger = MagicMock()
-        self.spider.stats_dict = {}
-
-    def get_meta(self):
-        item = {}
-        item['crawlid'] = "abc123"
-        item['appid'] = "myapp"
-        item['spiderid'] = "link"
-        item["attrs"] = {}
-        item["allowed_domains"] = ()
-        item["allow_regex"] = ()
-        item["deny_regex"] = ()
-        item["deny_extensions"] = None
-        item['curdepth'] = 0
-        item["maxdepth"] = 1
-        item['priority'] = 0
-        item['retry_times'] = 0
-        item['expires'] = 0
-
-        return item
-
-    def do_test(self, meta_object,
-                text, expected_raw, expected_requests):
-        request = Request(url='http://www.drudgereport.com',
-                          meta=meta_object)
-        response = HtmlResponse('drudge.url', body=text, request=request)
-
-        raw_item_count = 0
-        request_count = 0
-
-        for x in self.spider.parse(response):
-            if isinstance(x, RawResponseItem):
-                raw_item_count = raw_item_count + 1
-            elif isinstance(x, Request):
-                request_count = request_count + 1
-
-        self.assertEqual(raw_item_count, expected_raw)
-        self.assertEqual(request_count, expected_requests)
-
-    def test_link_spider_parse(self):
-        with open('tests/drudge.html', 'r') as file:
-            text = file.read()
-
-            curr_meta = self.get_meta()
-
-            # test too deep
-            curr_meta['maxdepth'] = 1
-            curr_meta['curdepth'] = 1
-            self.do_test(curr_meta, text, 1, 0)
-
-            # test raw depth recursion
-            curr_meta['curdepth'] = 0
-            self.do_test(curr_meta, text, 1, 391)
-
-            # test allowed_domains filter
-            curr_meta['allowed_domains'] = ['foxnews.com']
-            self.do_test(curr_meta, text, 1, 4)
-
-            # test allow regex filter
-            curr_meta['allowed_domains'] = ()
-            curr_meta['allow_regex'] = ['.*shock.*']
-            self.do_test(curr_meta, text, 1, 1)
-
-            # test deny regex filter
-            curr_meta['allow_regex'] = ()
-            curr_meta['deny_regex'] = ['.*.com.*']
-            self.do_test(curr_meta, text, 1, 43)
-
-            # test deny_extensions filter
-            curr_meta['deny_regex'] = ()
-            # no pages that end in .html
-            curr_meta['deny_extensions'] = ['html']
-            self.do_test(curr_meta, text, 1, 329)
-
-if __name__ == '__main__':
-    unittest.main()
